pyproject.toml
[tool.poetry]
name = "test-pets"
version = "0.1.0"
description = ""
authors = ["nvladimirov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pytest = "^8.3.4"
allure-pytest = "^2.13.5"
requests = "^2.32.3"
assertpy = "^1.1"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
